# docs/openapi/schema.yaml
openapi: 3.0.3
info:
  title: KenzieShoesAPI
  version: 0.0.1
  description: Kenzie shoes api is about to create an e-commerce shoe store
paths:
...
  /api/users/buyorders/:
    get:
      operationId: users_buyorders_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderList'
          description: ''
...
  /api/users/orders/:
    post:
      operationId: users_orders_create
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Order'
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: ''
  /api/users/orders/{id}/:
    put:
      operationId: users_orders_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatus'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OrderStatus'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderStatus'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatus'
          description: ''
    patch:
      operationId: users_orders_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatus'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatus'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatus'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatus'
          description: ''
  /api/users/sellorders/:
    get:
      operationId: users_sellorders_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderList'
          description: ''
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        purchased_at:
          type: string
          format: date-time
          readOnly: true
        product:
          type: integer
          readOnly: true
          nullable: true
        quantity:
          type: integer
          readOnly: true
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        total_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        product_name:
          type: string
          readOnly: true
        product_category:
          type: string
          readOnly: true
        product_image:
          type: string
          format: uri
          readOnly: true
        seller:
          type: integer
          readOnly: true
          nullable: true
        buyed_at:
          type: string
          format: date-time
          readOnly: true
        user_buy:
          type: string
          format: email
          readOnly: true
        products:
          type: integer
          readOnly: true
      required:
      - buyed_at
      - id
      - product
      - product_category
      - product_image
      - product_name
      - products
      - purchased_at
      - quantity
      - seller
      - status
      - total_price
      - unit_price
      - user_buy
    OrderStatus:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        purchased_at:
          type: string
          format: date-time
          readOnly: true
        product:
          type: integer
          readOnly: true
          nullable: true
        quantity:
          type: integer
          readOnly: true
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        total_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        product_name:
          type: string
          readOnly: true
        product_category:
          type: string
          readOnly: true
        product_image:
          type: string
          format: uri
          readOnly: true
        seller:
          type: integer
          readOnly: true
          nullable: true
        buyed_at:
          type: string
          format: date-time
          readOnly: true
        user_buy:
          type: string
          format: email
          readOnly: true
        products:
          type: integer
          readOnly: true
      required:
      - buyed_at
      - id
      - product
      - product_category
      - product_image
      - product_name
      - products
      - purchased_at
      - quantity
      - seller
      - total_price
      - unit_price
      - user_buy
    PaginatedOrderList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Order'
...
    PatchedOrderStatus:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        purchased_at:
          type: string
          format: date-time
          readOnly: true
        product:
          type: integer
          readOnly: true
          nullable: true
        quantity:
          type: integer
          readOnly: true
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        total_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        product_name:
          type: string
          readOnly: true
        product_category:
          type: string
          readOnly: true
        product_image:
          type: string
          format: uri
          readOnly: true
        seller:
          type: integer
          readOnly: true
          nullable: true
        buyed_at:
          type: string
          format: date-time
          readOnly: true
        user_buy:
          type: string
          format: email
          readOnly: true
        products:
          type: integer
          readOnly: true
...
    StatusEnum:
      enum:
      - Realizado
      - Em andamento
      - Entregue
      type: string
      description: |-
        * `Realizado` - Realizado
        * `Em andamento` - Andamento
        * `Entregue` - Entregue
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT